refactor: modularize app architecture
