+++ index.html
@@ -627,6 +627,10 @@
     <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
     <script src="assets/js/storage.js"></script>
     <script src="assets/js/categories.js"></script>
+    <script src="assets/js/constants.js"></script>
+    <script src="assets/js/data-service.js"></script>
+    <script src="assets/js/charts.js"></script>
+    <script src="assets/js/ui.js"></script>
     <script src="assets/js/main.js"></script>
   </body>
 </html>
